feat: 과제 관련 클래스 및 도메인 분리 (homework, quiz, video, zoom)
