--- scripts/side-panel.js
@@ -1,3 +1,0 @@
-export default function sidePanel() {
-  console.log('side-panel.js loaded');
-}
