+++ scripts/type/assignment.type.js
@@ -1,0 +1,11 @@
+/**
+ * @enum {{ HOMEWORK: string, QUIZ: string, VIDEO: string, ZOOM: string }} ASSIGNMENT_TYPE
+ */
+const ASSIGNMENT_TYPE = {
+  HOMEWORK: 'homework',
+  QUIZ: 'quiz',
+  VIDEO: 'video',
+  ZOOM: 'zoom',
+};
+
+export default ASSIGNMENT_TYPE;
